requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4        
certifi==2020.6.20       
chardet==3.0.4
charset-normalizer==2.0.9
comtypes==1.1.10
contextvars==2.4
cssselect==1.1.0
cssutils==2.3.0
EasyProcess==0.3
entrypoint2==0.2.4       
geographiclib==1.52
geopy==2.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.16
importlib-metadata==4.8.3
keyring==23.4.0
lxml==4.7.1
mss==6.1.0
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
Pillow==8.4.0
playsound==1.3.0
premailer==3.10.0
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pynput==1.7.5
pyparsing==3.0.6
pypiwin32==223
pyscreenshot==3.0
pyttsx3==2.90
pywin32==302
pywin32-ctypes==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wikipedia==1.4.0
wincertstore==0.2
WMI==1.5.1
yagmail==0.14.260
youtube-search==2.1.0
zipp==3.6.0